+++ src/main/java/com/java/JustThree/controller/BoardController.java
@@ -6,8 +6,8 @@ import com.java.JustThree.dto.board.response.GetBoardListResponse;
 import com.java.JustThree.dto.board.response.GetBoardOneResponse;
 import com.java.JustThree.dto.board.request.UpdateBoardRequest;
 import com.java.JustThree.exception.BoardNotFoundException;
-import com.java.JustThree.service.BoardReplyService;
-import com.java.JustThree.service.BoardService;
+import com.java.JustThree.service.board.BoardReplyService;
+import com.java.JustThree.service.board.BoardService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
