+++ src/main/java/com/java/JustThree/service/board/BoardReplyService.java
@@ -1,4 +1,4 @@
-package com.java.JustThree.service;
+package com.java.JustThree.service.board;
 
 import com.java.JustThree.domain.Board;
 import com.java.JustThree.domain.BoardReply;
